rdbms: Fix test to not rely on deprecated class aliases
Restore unrelated test fix from 46eabe275c, which got reverted
in 0b43c49465.

The DatabaseFactory::create() function takes $type which is either
a built-in type (mysql, sqlite, postgres) or theoretically the suffix
of a database extension class for a custom type.

"mysqli" is not a valid type, the type is "mysql". This was passing
because DatabaseMysqli still exists as a class alias for the namespaced
Rdbms class.

Bug: T309418
Change-Id: Ic05a4b3d4cfe539e11df3dc83dfe459c184d1332