Prevent new accounts from using = in their usernames
Equals signs have special meaning in wikitext, they're used in template
parameters and headings. Allowing them in usernames complicates things
as it's not possible to use them in wikitext without escaping them.

This only prevents usage in new accounts, so it doesn't stop the problem
entirely, but it stops it from getting worse.

Bug: T254045
Change-Id: I07c0af2c5e44ff67fc0bc6c1ae1d4ce7255aa0cd