Dynamically update template list for live preview
This adds better handling for the template list under the edit
form, making sure that it updates on live preview and uses the
correct messages. It also sorts the templates, and applies CSS
classes for redlinks and redirects etc.

A 2nd API call is required for this, to fetch the template page
protection status so that edit links can be added correctly.
This incidentally also gets redirect status, which is good, but
it does make for more network traffic.

Bug: T321032
Bug: T327354
Change-Id: I61c4080bc432a0dae03083a027890e7de786898c