parser: Use the actual revision timestamp of the page for messages
This patch aims to fix T320338#8334752, any message parsing that
happened when (pre)viewing a page should use the actual timestamp of
the last saved revision. The title is the same as the current page
only for context.

For the case reported in T320338#8334752, edit notice should indicate
the time since the last edit properly.

Bug: T320338
Change-Id: I2a5d44087e74efa4a7e2bb1b75036506e8e3823a