Switch search results padding-bottom to margin-bottom
It was intentionally padding rather than bottom, to allow
stacking it on top of the global li margin-bottom of 0.1em.

That's pointless, though: that 0.1em is negligible compared
to the 24px we're adding, and the context is very, very
different from the regular text list-items this is meant to
control - there's absolutely no need to also keep that one
around. Overriding that margin makes more sense here, since
it gives us better control over what exactly we want this
to look like.

And in one specific case, for Extension:SearchVue, it is
impractical for the whitespace between search results to be
expanded via padding instead of margin: it inserts a
full-height node within the search result, and that ends up
filling the whitespace-included height, looking unbalanced.

Bug: T310281
Change-Id: If2ea567ad77c373d2075c2a2cb97dff104fe3876