SpecialAllPages: Remove object check after Title::newFromRow
Title::newFromRow always return a Title object and never returns null.
In case of an invalid title like '<' or an invalid namespace in the
database the title Object contains the values from the database.

This change removes the dead code because $t is never falsey.

Also use Title::isRedirect() to identifiy redirects.

Change-Id: Ide18f3092520831112be40b16d231c58f859a18a