Add tests for taint-check
Make sure that taint-check knows about these methods, especially as we
move the taintedness info from the plugin to annotations on the methods.

Some tests are currently disabled because they would fail. These will
be enabled in subsequent commits while adding annotations to the
relevant methods.

Note that the disabling needs to be done by making the suppression
annotation invalid (done by prepending 'xxx' in this patch), because
commenting out the code or adding a method-wide @suppress won't work.

Bug: T321806
Change-Id: I8b89a22a5c23a2ab25329bcb06c673168d24683d