api: Fix order by on list=allimages&aisort=timestamp for pagination
The order by on list=allimages&aisort=timestamp is only timestamp,
which is not unique, include the name to get unique pagination.
The continue parameter already working like that.

With d9534a5 the img_name was added to ORDER BY using addWhereRange,
but that is overridden by the explicit ORDER BY added later,
Remove the explicit ORDER BY to use the ORDER BY from
addTimestampWhereRange/addWhereRange

Change-Id: Id4783c37372847841ac86526df28aeed8b8eb0f7