ResourceLoader: Add 1min grace via stale-while-revalidate Cache-Control
Only enable it if maxage is above 1 minute in the first place, e.g.
not for debug/error responses or when otherwise having configured
things during development with a very low maxage.

This gives blocking resources like the startup manifest and stylesheets
a chance to never block page rendering other than during the first
view of a browsing session. Once a browsing session reaches the 5min
mark, instead of blocking the next page load, it could then be renewed
in the background if the pageview starts during the 6th minute.

CDNs and web proxies known to support the stale-while-revalidate
directive:
* Varnish 4.1.0
* NGINX 1.11.10
* Squid 2.7

Browsers known to support it:
* Firefox 68 (2019)
* Chrome 75 (2019)

Bug: T132418
Change-Id: Ifae1f09722be4abf1c4dfe895122a3503010a422