pager: Remove unused AlphabeticPager::getOrderTypeMessages() support
This was introduced in March 2008 with r32228 (ad178edb80) for use in
the CategoryPager of Special:Categories, but almost immediately
removed again from there, on the same day, in r32259 (bdf9431795)
because it wasn't actually supported by the underlying DB index.

There has been no other use of it to my knowledge. Removing without
deprecation as such. Note that (theoretical) use will not result in
visible breakage as the public stability is limited to implementing
(not calling) the method. Calling is private and the only caller
is removed in this commit.

The call to setExtra() is removed as well as it is now always empty,
which is identical to the default in PagerNavigationBuilder, and there
are no other calls to setExtra before this point that this was removing
or replacing data for.

Bug: T321551
Change-Id: I161dc0159e4372e3478341ee3fbea13b723d9fc1