Make RequestContext::sanitizeLangCode() accept null
I think it's within the meaning of the method to sanitize the type of
the input. Code review shows that it's almost always called with
technically nullable input, so it's convenient to deal with nulls at
this single place.

The linked bug is a PHP 8.1 deprecation warning due to a test which set
up a StaticUserOptionsLookup without a language option.
StaticUserOptionsLookup does not fall back to the site defaults, so the
language was null, causing an error from strtolower().

Bug: T322099
Change-Id: I6dc61476c6869a7648a67be79a4835a5ac24fa92
(cherry picked from commit d47ca15928c930d99b8669a4ce731f65e20b134f)