filerepo: Fix timestamps being incompatible with some databases
The confusion is because we typically work with so called TS_MW
timestamps in all code. These are strings with 14 digits,
"YYYYMMDDHHMMSS". These work in some databases, but not all. That's
why the $dbw->timetamp() method exists, to convert to the format
accepted by the currently used database.

This code missed some of these calls.

I assume this was never found because we never tested carefully with
other databases, and still rarely do.

I found this error after a "check experimental" in Ie39bd0e,
triggered by a FileImporter test.

Note that there are already some $dbw->timestamp() calls in this
code, exactly like the ones I add.

Change-Id: I29399e7fda775ecb2d70edb50a0fe18d1f48ff3f