ApiQueryUserContribs: Remove unused continuation modes
When using the 'userids' or 'user' parameters, $this->orderBy would
previously be set to (respectively) 'id' or 'name'.

This was confusing, because:

* If more than 1 user ID/name was requested, it was immediately
  overwritten to 'actor' by the code below.

* If only 1 user ID/name was requested, it was not used for building
  either the continuation string or the database query.

Set it to 'actor', and remove code generating and parsing the unused
continuation modes (note that 'name' is still used with other
parameters). Also rename a variable called `$fromName` that actually
always contained an actor ID and not a user name (resolving a FIXME).

Follow-up to c29909e59fd89d9fc65a299687ad2aa2df995acc, which removed
the code paths that used 'id' or 'name' continuation modes for paging
by user_id and user_name.

Change-Id: I9570292beaa7dca0dbc165ab319b9eef42712efa