Fix off-by-1 error in Preprocessor_Hash
Here is an example wikitext to play around with:
a
<!----> <!---> <!--->
b

a
<!----> <!---> <!--->b

https://www.mediawiki.org/w/api.php?action=parse&text=a%0A%3C!----%3E%20%3C!---%3E%20%3C!---%3E%0Ab%0A%0Aa%0A%3C!----%3E%20%3C!---%3E%20%3C!---%3Eb&prop=parsetree

Notice how the comments are parsed different in the two situations.
<comment>&lt;!----&gt; </comment><comment>&lt;!---&gt; </comment><comment>&lt;!---&gt;\n</comment>
<comment>&lt;!----&gt;</comment> <comment>&lt;!---&gt; &lt;!---&gt;</comment>

The first one is wrong because <!---> is not an HTML comment. There
is one dash missing. It should be parsed as <!-- followed by the
content "-> <!-" followed by the closing -->, just as in the second
example.

Change-Id: Iaf8e52c5d41ca3f28890e8b66cfdf1d86126c357