SECURITY: Make sqlite DB files not world readable
Ensure that SQLite DB files are created as 0600. Previously they
were created 0644 (world-readable). This was somewhat mitigated by
the fact that if MediaWiki had to create the parent directory, it
would be created as 0700. However if the parent directory already
existed and had loose permissions, than anyone on the server
could read the DB file and all the private data contained therein.

This only affects newly created DB files during the installation
process. It does not adjust previously created DB files.

Bug: T322637
Change-Id: I167ec1b5cd9a7e96919886fd20917b5733d59fc7