Treat search result suggestions like clicks
If the user selected a search result with the keyboard, treat it like a
click and directly navigate to the result’s URL, skipping the usual form
submit and its associated tracking. This is necessary to keep the
tracking consistent with the new Vue search, following the similar
changes made there by Codex change I49232bb11e.

If we don’t have a URL, because the wrapAsLink config was set to false,
we have no choice but to let the form submit go through; avoid tracking
an extra click event in that case.

Bug: T317682
Change-Id: Iea770d734a8b5e56f9199b09bb79db4df1127d2e