installer: Use WAL journal mode when creating main Sqlite DB
SQlite has multiple journal modes. Previously we used "delete"
for the main DB, and "wal" for the cache DB.

WAL mode has much better performance when people are reading
and writing to the DB at the same time. The downside is that
WAL mode is not safe if the DB is on a network share, and
processes on different hosts are writing to the DB at the same
time. It also has somewhat reduced guaruntees of durability
in the event of a power failure (but it does guaruntee no corruption). If the DB had no write load, delete mode is slightly faster.

We already use WAL on the cache sqlite DBs, so presumably we
are ok with its downsides. Since we are already using it,
we might as well fully reap the performance benefits. See
the bug for some quick and dirty benchmarking.

Bug: T322633
Change-Id: Idb432f5bde7e10cea064b4d2fca96216cd597b77