Phan: parse only directly required composer dependencies
Indirect dependencies are typically needed only by the direct
dependencies, which are not analyzed.

Dev dependencies are needed to run Phan, but are not needed in the code
Phan is analyzing.

So it turns out to be fairly straightforward to eliminate unnecessary
dependencies from the Phan parse stage, reducing memory usage by about
40% and time usage by about 20%. This means I can run Phan on my laptop
without first closing PHPStorm, which is convenient.

As a matter of policy, you could argue that the list of allowed indirect
dependencies here should be explicitly declared by moving them to
composer.json.

Change-Id: I987c3b22bafd1b4332fdeaf40ec3e20d35a3d929