Treat search result suggestions like clicks (v2)
If the user selected a search result with the keyboard, treat it like a
click and directly navigate to the result’s URL, skipping the usual form
submit and its associated tracking. This is necessary to keep the
tracking consistent with the new Vue search, following the similar
changes made there by Codex change I49232bb11e.

If we don’t have a URL, because the wrapAsLink config was set to false,
we have no choice but to let the form submit go through; avoid tracking
an extra click event in that case.

This is an identical resubmit of change Iea770d734a (5d5ee8014f), which
only had to be reverted to wait for a Codex update (done in I58becd2b5a).

This reverts commit 8ad9ad1bd25de3a8700bc7d9de9d668eef59a569.

Bug: T317682
Change-Id: I6b2f466c4a21c824d3acc62d42785c590b4c7726