Disable php in .htaccess of images directory as a hardening measure.
All our docs strongly reccomend doing this. There is even a prompt
in the installer. Not all webservers listen to this of course, but
it won't hurt things that ignore it.

The general idea is that there should be no directory that is
both writable and executable at the same time by the webserver.
images must be writable, so we should turn off php so its not executable.

Change-Id: Ic03cee12845a56a0f4f7e356493eb0f446ccf34c
(cherry picked from commit def09fd9ebe1da051709f816d47c8313be5b514d)