SpecialUserrights: Replace deprecated User::get/add/removeGroups
For the function userCanChangeRights/doSaveUserGroups the manager
instance of the special page is not usable, because that is only set
from the result of fetchUser(), but both functions acting as
backend-function and called without submitting the special page, which
would set mFetchUser after fetchUser().
When mFetchUser is not set, the manager instance is also not set.
Both function needs to create an own manager to handle that case.

Bug: T275148
Change-Id: Iade0cc0e75870ec64dd5f9e7a6929f031f9b5629