specialpage: Silence known violation unsafe RequestContext changes
RequestContext warns against unsafe changes to RequestContext because
context objects are injected and widely used in hooks, thus once the
init phase is over, changing it would have unpredictable and partial
effects on callers e.g. not agreeing on what the current request is
for.

The one case that we found where this is "intended", unfortunately,
is SpecialPage transclusion by the parser where we effectively try
to run an embedded request for something else and return the result.

While the SpecialPage object receives the context via executePath()
$page->setContext(), there are still some extensions out there that
through hooks called from a SpecialPage subclass inspect global state.
This is well-known and is why SpecialPageFactory also temporarily replaces
various global variables and modifies the "main" context object.

Acknowledge this known violation with a silencing so that we can
move on with catching bugs and preventing regressions everywhere
else.

Bug: T323184
Change-Id: Ibe4537e3f161a1da59c5237b82dfbaf207452808