rdbms: make doUpsert() handle SET clauses with TIMESTAMPTZ
For postgres, work around the lack of implicit type casting for assignments
like "column=(SELECT ...)" within the SET clause of UPDATE statements. While
implicit casting applies to "INSERT...VALUES", it does not apply to INSERT,
nor UPDATE, when the rows are selected from a Common Table Expression defined
using VALUES(). Query the table schema and explicitly set the types in this
case.

Change-Id: I0affc3a794dd640c5fedf765a5058c38dc353471