MediaWiki: Temp silence FR-induced clearActionName warnings
There's a few remaining for the following conditions are met:
1. The wiki has FlaggedRevs installed.
2. The URL is to a wikipage redirect.
3. The destination page uses FR and has unreviewed changes pending.

In this case, FlaggedRevs uses onInitializeArticleMaybeRedirect
and calls RequestContext::getActionName which is then invalidated
due to changing the Title/WikiPage/ContentHandler from which actions
are determined.

This will be fixed in FlaggedRevs soon, but for now silence it while
we take that on carefully.

Bug: T323254
Change-Id: If86476da7bc76be49bad0780dc63d2255080aaa2