Make RestStructureTest run in IDE
PHPUnit tests will fail to run in the IDE if they access the service
container.

Note that the structure test will not include routes registered by
extensions when run without the phpunit wrapper: data providers are
executed before config is loaded, so extensions do not have a chance to
register.

Change-Id: If9d408456686b408062204bec4b41c95f60ab775