HistoryPager: Prettify URL params of form submissions without JS
Split the form that was used for both comparing revisions and changing
visibility / editing tags into two by using `<input form="…">`
attributes. This way irrelevant parameters will not be included in
the form submission, and we won't need to clean them up in JS.

The forms still work correctly in browsers that don't support the
attribute (among browsers we otherwise support, this is only IE 11:
https://caniuse.com/form-attribute). The irrelevant parameters are
included with all form submissions, but I think that's acceptable.

Bug: T323338
Change-Id: Ia1157dfba22e2093f294c8599ebe3d7782a6abb3