password: Reduce time cost of password unit tests
Remove Argon2PasswordTest::testPartialConfig(). It's irreparably slow
and doesn't seem to test anything useful, just that the defaults are
the defaults. It was added in PS5 of Ifdf648f5d8a734a663e6 as part of a
change which removed explicit cost defaults from DefaultSettings.php.

Reduce time costs of some other tests.

The time to execute tests in this directory is reduced from ~1.3s to
~0.5s.

Bug: T225730
Change-Id: I30b846b5364ad2030e2547da6c58db13eaf401f2