rdbms: clarify "waitForPos" logic in LoadBalancer
Rename lazyLoadReplicationPositions() and doWait(). The later now
handles the logic of checking if a position is set or if the reader
index is that of the primary DB.

Narrow the job of reallyOpenConnection() by removing "waitForPos"
loading logic.  It now only deals with initializing a connection and
not the logic that goes along with tracking/pooled connection handles.
Note that getReaderIndex() still triggers primary position loading and
was the only method that automatically triggered logic to actually wait
on the primary position.

Add more comments to getReaderIndex() and optimize call order slightly.

Minor cleanups to related code comments.

Change-Id: I685b1f7946aadb5463e8870edda4340be3fc4ae2