HtmlOutputRendererHelper: Add edit flavor
This adds support for an "edit" flavor to HtmlOutputRendererhelper,
which causes data-parsoid attributes to be present in the HTML output.

In addition, this adds the ability to request HTML output using a
specific profile versionThis adds support for an "edit" flavor to
HtmlOutputRendererhelper, which causes data-parsoid attributes to be
present in the HTML output.

In addition, this adds the ability to request HTML output using a
specific profile version..

Change-Id: I5c21e0358da6a8915ddfe574fb2d13e469f625c5