Use SelectQueryBuilder in BacklinkCache
Mostly replace the traditional / ANSI-89 JOINs in this class (i.e.,
select from two tables, empty $joinConds, actual join condition in
$conds) with proper JOINs, and also mostly replace the “kind of janky”
$conds filter in queryLinks(): instead, initQueryBuilderForTable(), the
new replacement for getConditions(), now knows whether the page table
needs to be joined or not.

However, the hook still needs to be supported (there is one extension in
codesearch that actually uses it and seems to be maintained), and in
that hook we don’t know which part of the $conds are JOIN conditions; in
this case, we have to keep the “kind of janky” filter, and also join the
page table using table() instead of join().

Bug: T311866
Change-Id: Ic49bf5d8fd136d296752e8807f6032bfae74dcdf