ParserCache: fix metrics keys
Keep metrics keys indicating the reason for saving to the cache
separate from keys that indicate the outcome of trying to save.
This is needed to fix Grafana dashboards that evaluate the save_* key
prefix. Such dashboards have started to count most operations twice, once
for save_success and once for save_reason_*

NOTE: We were indiscriminately replacing "." with "_" in all keys.
Per this patch, all established keys will keep using their old format
with a "_", while the new keys that indicate the reaons will use a ".".
It would be nicer to use "." in the old keys as well, but that would
break existing dashboards.

See https://grafana.wikimedia.org/d/000000106/parser-cache?orgId=1&from=1669815989493&to=1669988789493&viewPanel=14

Bug: T324216
Change-Id: I7aa7ba98bf26a17969939eb1366d7c474c469431