search: Set file_text to null when not available
file_text has previously been set to false or an empty array,
depending on context, when it wasn't available. As part of normalizing
the set of types used in the search index default the value to null
and only set it if the media handler is able to extract text content
from it.

Setting the value to null when not available is done to clear out
historical false/empty array values in storage. Perhaps we need to
think more in the future about if/when default values should be
provided to search updates, everything is a bit ad-hoc today.

Bug: T322327
Change-Id: I1367154b17d9e69c9373e7efee384838aa3b51e8